# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Бухгалтер API
  description: |
    API приложения «Бухгалтер» для учёта личных финансов.

    Ответы с ошибками (`4xx`, `5xx`) используют объект `error` с полями `code`, `message` и опционально
    **`field`** — имя поля тела запроса для подсветки в форме UI (например `monthly_payment`, `term_months`).
    Схема: `ErrorResponse` → `error.field`.

    **Серверный кеш (in-memory):** успешные ответы `GET` кешируются в процессе сервера (справочники — 5 мин,
    остальное — 1 мин); при любой мутации кеш пользователя сбрасывается.
    **Клиент (web / Android):** SWR ref-cache в `localStorage` для `GET /api/v1/*` (кроме health, export, preview, version, `GET /setup/status`).
    Подробнее — [ui-api-cache.md](../ui-api-cache.md).

    **Счета (UI):** на `/accounts` меню «⋯» вызывает inline-редактирование (`PUT /accounts/{id}`),
    `PUT /accounts/{id}/credit-limit` (кредитные карты), `POST /accounts/{id}/primary`,
    `POST /accounts/{id}/archive`, `DELETE /accounts/{id}`.
    На `/accounts/{id}` — та же форма редактирования в шапке. Поля: название, банк, начальный баланс.
    Лимит кредитной карты меняется только через `PUT /accounts/{id}/credit-limit` (см. [ui-credit-cards.md](../ui-credit-cards.md)).
    В форме «Начальный баланс» — значение `initial_balance`, не текущий `balance_display` (см. [data-model.md](../data-model.md)).
    Подробнее — [ui-row-actions.md](../ui-row-actions.md).
  version: 1.5.0
servers:
  - url: /api/v1
paths:
...
  /setup:
    post:
      summary: Начальная настройка (один раз)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - admin_login
                - admin_display_name
                - admin_password
                - admin_password_confirm
              properties:
                admin_login:
                  type: string
                  minLength: 3
                  maxLength: 32
                  description: Логин администратора
                admin_display_name:
                  type: string
                  minLength: 1
                  maxLength: 64
                  description: Имя администратора (display_name)
                admin_password:
                  type: string
                  minLength: 8
                  description: Минимум 8 символов, хотя бы одна буква и одна цифра, не совпадает с admin_login
                admin_password_confirm:
                  type: string
                  minLength: 8
                  description: Должен совпадать с admin_password
                registration_enabled:
                  type: boolean
                  default: false
                external_url:
                  type: string
                  description: Опционально; URL за reverse proxy (http/https)
      responses:
        '201':
          description: Настройка завершена
        '400':
          description: Ошибка валидации (пароли не совпадают, пустое имя и т.д.)
        '409':
          description: Настройка уже выполнена
  /setup/restore:
    post:
      summary: Восстановить БД до завершения настройки
      description: |
        Доступно до завершения первичной настройки.
        Multipart: поле `file` (`.db`) и `confirm=RESTORE`.
        После восстановления возвращает `configured`:
        - `true` — восстановленная БД уже настроена, можно переходить на `/login`;
        - `false` — завершите первичную настройку на `/setup`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, confirm]
              properties:
                file:
                  type: string
                  format: binary
                  description: Файл SQLite (.db)
                confirm:
                  type: string
                  enum: [RESTORE]
      responses:
        '200':
          description: Восстановление завершено
          content:
            application/json:
              schema:
                type: object
                required: [message, configured]
                properties:
                  message:
                    type: string
                    example: restore_complete
                  configured:
                    type: boolean
        '400':
          description: Некорректный multipart, confirm или файл
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                restore_confirm:
                  $ref: '#/components/examples/ApiErrorRestoreConfirm'
        '409':
          description: Инстанс уже настроен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                already_configured:
                  $ref: '#/components/examples/ApiErrorAlreadyConfigured'
        '500':
          description: Ошибка восстановления
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                internal:
                  $ref: '#/components/examples/ApiErrorInternal'
  /auth/login:
    post:
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успешный вход
        '401':
          description: Неверные учётные данные
        '403':
          description: Учётная запись на модерации или заблокирована (USER_PENDING_MODERATION, USER_BANNED)
        '429':
          description: Превышен лимит попыток
  /auth/logout:
    post:
      summary: Выход из системы
      security:
        - sessionCookie: []
      responses:
        '204':
          description: Выход выполнен
  /auth/register:
    post:
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
                - password_confirm
              properties:
                login:
                  type: string
                  minLength: 3
                  maxLength: 32
                password:
                  type: string
                  minLength: 8
                  description: Минимум 8 символов, хотя бы одна буква и одна цифра, не совпадает с login
                password_confirm:
                  type: string
                  minLength: 8
                  description: Должен совпадать с password
                display_name:
                  type: string
                  description: Опционально; по умолчанию — login
      responses:
        '201':
          description: Пользователь создан (статус pending, сессия не выдаётся)
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации (пароли не совпадают и т.д.)
        '403':
          description: Регистрация отключена
  /auth/request-password-reset:
    post:
      summary: Запросить сброс пароля
      description: |
        Публичный endpoint (без сессии). При существующем логине создаётся запись в
        `password_reset_requests` и уходит уведомление администраторам (`password_reset`).
        При неизвестном логине ответ тот же — без перечисления учёток.
        Rate limit: 5 запросов/мин с IP (`Retry-After: 60` при `429`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login]
              properties:
                login:
                  type: string
                  description: Логин пользователя
      responses:
        '204':
          description: Запрос принят (логин существует или нет — одинаковый ответ)
        '400':
          description: Пустой login или невалидный JSON (`VALIDATION_ERROR`)
        '429':
          description: Превышен лимит запросов с IP (`RATE_LIMITED`)
  /auth/verify:
    get:
      summary: Проверка токена
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Результат проверки
  /features:
    get:
      summary: Снимок флагов модулей
      description: |
        Инстанс-флаги опциональных модулей для UI (навигация, роуты).
        Требуется авторизация. При выключенном модуле соответствующий API отвечает 404 `FEATURE_DISABLED`.
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: Карта key → enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagsSnapshot'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                unauthorized:
                  $ref: '#/components/examples/ApiErrorUnauthorized'
  /auth/me:
    get:
      summary: Текущий пользователь
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Данные пользователя
  /user/settings:
    get:
      summary: Настройки пользователя
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Настройки
    put:
      summary: Обновить настройки
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Настройки обновлены
  /user/password:
    put:
      summary: Смена пароля
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_password
                - new_password
                - new_password_confirm
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
                  description: Минимум 8 символов, хотя бы одна буква и одна цифра, не совпадает с login
                new_password_confirm:
                  type: string
                  minLength: 8
                  description: Должен совпадать с new_password
      responses:
        '204':
          description: Пароль изменён
        '400':
          description: |
            Ошибка валидации. Код ответа `VALIDATION_ERROR`; тексты сообщений по ключам:
            `PASSWORDS_MISMATCH`, `INVALID_CURRENT_PASSWORD`, `PASSWORD_TOO_SHORT`,
            `PASSWORD_TOO_WEAK`, `PASSWORD_UNCHANGED` (новый пароль совпадает с текущим).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                password_unchanged:
                  $ref: '#/components/examples/ApiErrorPasswordUnchanged'
  /user/tokens:
    get:
      summary: Список API-токенов
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Список токенов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIToken'
    post:
      summary: Создать API-токен
      description: |
        По умолчанию срок действия — 30 дней. Бессрочный токен (`never_expires: true`) допустим,
        но не рекомендуется. Явная дата в `expires_at` (RFC3339) переопределяет срок по умолчанию.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APITokenCreateRequest'
      responses:
        '201':
          description: Токен создан (значение `token` показывается один раз)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APITokenCreated'
        '400':
          description: Ошибка валидации (`ERR_TOKEN_NAME_REQUIRED`, `ERR_TOKEN_EXPIRES`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                token_name_required:
                  $ref: '#/components/examples/ApiErrorTokenNameRequired'
                token_expires:
                  $ref: '#/components/examples/ApiErrorTokenExpires'
  /user/tokens/{id}:
    delete:
      summary: Отозвать API-токен
      description: Можно отозвать только собственный токен; чужой — `404 NOT_FOUND`.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Токен отозван
  /user/notifications:
    get:
      summary: Настройки уведомлений пользователя
      description: |
        Возвращает конфигурацию каналов (Telegram/MAX), триггеры и шаблоны.
        Секреты (`*_token`) в ответе не передаются.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Настройки уведомлений
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSettings'
    put:
      summary: Обновить настройки уведомлений и шаблоны
      description: |
        Поля в теле опциональны. Для токенов используется шифрование AES-256-GCM
        (ключ задаётся в админке, `system_settings.notification_secret_key`).
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettingsUpdate'
      responses:
        '200':
          description: Обновлённые настройки уведомлений
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSettings'
        '400':
          description: Ошибка валидации (неверный trigger_type, placeholder, длина шаблона и т.д.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                validation:
                  $ref: '#/components/examples/ApiErrorValidation'
  /user/notifications/test:
    post:
      summary: Отправить тестовое уведомление
      description: |
        Отправляет тестовый шаблон (`trigger_type=test`) в выбранный канал.
        Поддерживаются каналы `telegram` и `max`.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationTestRequest'
      responses:
        '200':
          description: Тестовое сообщение отправлено
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: sent
        '400':
          description: |
            Канал не настроен или некорректные параметры.
            Код `VALIDATION_ERROR`; сообщения: `ERR_CHANNEL` (неверный канал),
            `ERR_NOTIFICATION_CHANNEL` (канал не настроен), `ERR_INVALID_JSON`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                channel_not_configured:
                  $ref: '#/components/examples/ApiErrorNotificationChannel'
  /user/notifications/templates/preview:
    post:
      summary: Предпросмотр шаблона уведомления
      description: |
        Формирует текст шаблона с тестовыми данными без отправки в мессенджер.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreviewRequest'
      responses:
        '200':
          description: Результат предпросмотра
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreviewResponse'
        '400':
          description: Ошибка валидации шаблона
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                validation:
                  $ref: '#/components/examples/ApiErrorValidation'
  /user/notifications/templates/reset:
    post:
      summary: Сбросить пользовательские шаблоны уведомлений
      description: |
        Если `trigger_type` не передан — сбрасывает все пользовательские шаблоны.
        Если передан — только для указанного типа.
      security:
        - sessionCookie: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationResetRequest'
      responses:
        '200':
          description: Настройки уведомлений после сброса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSettings'
        '400':
          description: Некорректный trigger_type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                validation:
                  $ref: '#/components/examples/ApiErrorValidation'
...
  /admin/backups:
    get:
      summary: Список резервных копий
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Список файлов
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
...